% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-functions.R
\name{AUC}
\alias{AUC}
\title{AUC}
\usage{
AUC(pred, Y, ind = 1:length(Y), pivot = 0, binarise = TRUE, plot = TRUE)
}
\arguments{
\item{pred}{Predicted labels}

\item{Y}{True labels}

\item{ind}{Indices of instances to consider}

\item{pivot}{Pivot to split labels into positive and negative}

\item{binarise}{Indicator whether to binarise predictions prior to evaluation}

\item{plot}{Indicator whether to plot the AUC curve}
}
\value{
AUC
}
\description{
Compute the Area Under the ROC Curve for predicted labels
}
\author{
Fabian Schmich
}

