% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-functions.R
\name{cmn}
\alias{cmn}
\title{Class Mass Normalization (CMN) from Zhu et al., 2003}
\usage{
cmn(yhat, l, u)
}
\arguments{
\item{yhat}{Response for labeled (l) and unlabeld (u) genes}

\item{l}{Indices of labeled genes}

\item{u}{Indices of unlabeled genes}
}
\value{
Class normalized yhat
}
\description{
Class Mass Normalization (CMN) from Zhu et al., 2003
}
\author{
Fabian Schmich
}

