% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-functions.R
\name{cuthill_mckee}
\alias{cuthill_mckee}
\title{Cuthill McKee (CM) algorithm}
\usage{
cuthill_mckee(x)
}
\arguments{
\item{x}{Input matrix}
}
\value{
Band matrix
}
\description{
Transform sparse matrix into a band matrix
}
\author{
Fabian Schmich
}

