% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-methods.R
\name{fit}
\alias{fit}
\title{Fit \code{\linkS4class{netprioR}} model}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{A \code{\linkS4class{netprioR}} object}
}
\value{
A \code{\linkS4class{netprioR}} object with fitted model
}
\description{
Fit \code{\linkS4class{netprioR}} model
}
\author{
Fabian Schmich
}

