% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netprioR-functions.R
\name{gmrf}
\alias{gmrf}
\title{Sampling from a GMRF}
\usage{
gmrf(n, Q, M, xM, mu = 0, constr = list(A, e), use.inla = FALSE)
}
\arguments{
\item{n}{Number of samples to be drawn}

\item{Q}{Precision Matrix}

\item{M}{Indices of nodes to condition on}

\item{mu}{Mean of GMRF}

\item{A}{Constraint: Left side}

\item{e}{Constraint: Right side}
}
\value{
n samples from GMRF
}
\description{
From "GMRF Theory and Application" Alogirhtm 2.5 and Algorithm 2.6
}
\author{
Fabian Schmich
}

