% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netprioR-functions.R
\name{laplacian}
\alias{laplacian}
\title{Compute the laplacian matrix}
\usage{
laplacian(x)
}
\arguments{
\item{x}{Input matrix}
}
\value{
Laplacian matrix
}
\description{
Compute the laplacian matrix
}
\author{
Fabian Schmich
}

