% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netprioR-functions.R
\name{makegrid}
\alias{makegrid}
\title{Create a grid graph of n x m points}
\usage{
makegrid(n, m, dimnames = NULL)
}
\arguments{
\item{n}{Height of grid}

\item{m}{Width of grid}

\item{graph}{return graph instead of adjacency matrix}
}
\value{
Grid graph or adjacency matrix
}
\description{
Create a grid graph of n x m points
}
\author{
Fabian Schmich
}

