% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-methods.R
\docType{methods}
\name{weights}
\alias{ROC,netprioR-method}
\alias{ranks,netprioR-method}
\alias{weights}
\alias{weights,netprioR-method}
\title{\code{\linkS4class{netprioR}} accessors}
\usage{
weights(object, ...)

\S4method{weights}{netprioR}(object)

\S4method{ranks}{netprioR}(object)

\S4method{ROC}{netprioR}(object, true.labels, plot = FALSE)
}
\arguments{
\item{object}{A \code{\linkS4class{netprioR}} object}

\item{true.labels}{True full set of underlying labels}

\item{plot}{Flag whether to plot the AUC curve}
}
\value{
Estimated network weights, prioritisation ranks or ROC curve
}
\description{
Retrieve network weights, prioritisation ranks and ROC curves
}
\author{
Fabian Schmich
}

