% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-functions.R
\name{norm_kern}
\alias{norm_kern}
\title{Normalise kernel}
\usage{
norm_kern(x)
}
\arguments{
\item{x}{kernel}
}
\value{
Normalised kernel
}
\description{
adopted from GeneMania, Mostafavi et al, 2009
}
\author{
Fabian Schmich
}

