% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netprioR-functions.R
\name{pmat}
\alias{pmat}
\title{Construct precision matrix for GMRF}
\usage{
pmat(genes, interactions, label.priors, theta)
}
\arguments{
\item{theta}{Parameter to tune label bias}

\item{V}{Vector of (named) vertices}

\item{E}{Matrix of similarities / affinities, i.e. edges between vertices}

\item{labels}{String sequence of labels for vertices}
}
\value{
Precision matrix Q
}
\description{
Construct precision matrix for GMRF
}
\author{
Fabian Schmich
}

