% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-methods.R
\docType{methods}
\name{ranks}
\alias{ranks}
\alias{ranks,netprioR-method}
\title{Retrieve ranked prioritisation list}
\usage{
ranks(object)

\S4method{ranks}{netprioR}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{netprioR}} object}
}
\value{
Ranked list of prioritised genes
}
\description{
Retrieve ranked prioritisation list
}
\examples{
data(simulation)
ranks(simulation$model)
}
\author{
Fabian Schmich
}

