% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-methods.R
\docType{methods}
\name{ROC}
\alias{ROC}
\alias{ROC,netprioR-method}
\title{Compute ROC curve from netprioR model and true labels}
\usage{
ROC(object, ...)

\S4method{ROC}{netprioR}(object, true.labels, plot = FALSE, ...)
}
\arguments{
\item{object}{A \code{\linkS4class{netprioR}} object}

\item{...}{Additional arguments}

\item{true.labels}{True full set of underlying labels}

\item{plot}{Flag whether to plot the AUC curve}
}
\value{
ROC curve with AUC
}
\description{
Compute ROC curve from netprioR model and true labels
}
\examples{
data(simulation)
ROC(simulation$model, true.labels = simulation$labels.true)
}
\author{
Fabian Schmich
}

