% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-functions.R
\name{roccurve}
\alias{roccurve}
\title{Compute ROC curves from predicted and true labels}
\usage{
roccurve(labels.pred, labels.true, plot = TRUE, ...)
}
\arguments{
\item{labels.pred}{Predicted labels}

\item{labels.true}{True labels}
}
\value{
ROC object
}
\description{
Compute ROC curves from predicted and true labels
}
\author{
Fabian Schmich
}

