% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-simulation.R
\name{simulate_labels}
\alias{simulate_labels}
\title{Simulate labels}
\usage{
simulate_labels(values, sizes, nobs)
}
\arguments{
\item{values}{Vector of labels for groups}

\item{sizes}{Vector of group sizes}

\item{nobs}{Vector of number of observed labels per group}
}
\value{
List of Y, Yobs and indices for labeled instances
}
\description{
Simulate labels
}
\examples{
labels <- simulate_labels(values = c("Positive", "Negative"), 
sizes = c(10, 10), 
nobs = c(5, 5))
}
\author{
Fabian Schmich
}

