% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-simulation.R
\name{simulate_network_random}
\alias{simulate_network_random}
\title{Simulate random networks with predefined number of members for each
of the two groups and the number of neighbours for each node}
\usage{
simulate_network_random(nmemb, nnei = 1)
}
\arguments{
\item{nmemb}{Vector of number of members for each group}

\item{nnei}{Number of neighbours for each node}
}
\value{
Adjacency matrix of graph
}
\description{
Simulate random networks with predefined number of members for each
of the two groups and the number of neighbours for each node
}
\examples{
network <- simulate_network_random(nmemb = c(10, 10), nnei = 1)
}
\author{
Fabian Schmich
}

