% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-simulation.R
\name{simulate_network_scalefree}
\alias{simulate_network_scalefree}
\title{Simulate scalefree networks}
\usage{
simulate_network_scalefree(nmemb, pclus = 1)
}
\arguments{
\item{nmemb}{Vector of numbers of members per group}

\item{pclus}{Scalar in [0, 1] determining how strictly distinct groups are}
}
\value{
Adjacency matrix
}
\description{
Simulate scale free networks for predefined number of members for each of
two groups and a parameter pclus that determines how strictly distinct the groups
are
}
\examples{
network <- simulate_network_scalefree(nmemb = c(10, 10), pclus = 0.8)
}
\author{
Fabian Schmich
}

