% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-datasimulation.R
\name{simulate_phenotype}
\alias{simulate_phenotype}
\title{Simulate phenotypes correlated to labels pivoted into two groups}
\usage{
simulate_phenotype(Y, meandiff, sd, pivot = 0)
}
\arguments{
\item{Y}{Vector of labels}

\item{meandiff}{difference of means between positive and negative groups}

\item{sd}{Standard deviation of the phenotype}

\item{pivot}{Pivot point to split labels into groups}
}
\value{
Simulated phenotype
}
\description{
Simulate phenotypes correlated to labels pivoted into two groups
}
\author{
Fabian Schmich
}

