% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-package.R
\docType{data}
\name{simulation}
\alias{simulation}
\title{Example data: Simulated networks, phenotypes and labels for N = 1000 genes}
\usage{
data(simulation)
}
\description{
The data set contains simulated data for N = 1000 genes and P = 1 (univariate) phenotypes. 
The list of networks contains 2 low noise networks and two high noise networks. The class
labels are "Positive" and "Negative".
}
\details{
The code used to simluate the data can be found in system.file("example", "data_simulation.R", package = "netprioR")
}

