% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-methods.R
\docType{methods}
\name{weights}
\alias{weights}
\alias{weights,netprioR-method}
\title{Retrieve network weights}
\usage{
weights(object, ...)

\S4method{weights}{netprioR}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{netprioR}} object}

\item{...}{Additional arguments}
}
\value{
Estimated network weights
}
\description{
Retrieve network weights
}
\examples{
data(simulation)
weights(simulation$model)
}
\author{
Fabian Schmich
}

