% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-functions.R
\name{bandwidth}
\alias{bandwidth}
\title{bandwidth}
\usage{
bandwidth(x)
}
\arguments{
\item{x}{Inpute matrix}
}
\value{
Bandwidth
}
\description{
Compute the bandwidth of a matrix
}
\author{
Fabian Schmich
}

