% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-functions.R
\name{conjugate_gradient}
\alias{conjugate_gradient}
\title{Conjugate Gradient Solver}
\usage{
conjugate_gradient(A, b, x0 = rep(0, ncol(A)), threshold = 1e-15,
  verbose = FALSE)
}
\arguments{
\item{A}{Matrix}

\item{b}{Coefficients}

\item{x0}{Starting solution}

\item{threshold}{Termination threshold}

\item{verbose}{Show iterative progress}
}
\value{
Solution for equation system
}
\description{
Solves linear equation systems iteratively
}
\author{
Fabian Schmich
}

