% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-methods.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,netprioR-method}
\title{Fit \code{\linkS4class{netprioR}} model}
\usage{
fit(object, ...)

\S4method{fit}{netprioR}(object, refit = FALSE, ...)
}
\arguments{
\item{object}{A \code{\linkS4class{netprioR}} object}

\item{...}{Additional arguments}

\item{refit}{Flag whether to overwrite existing fit}
}
\value{
A \code{\linkS4class{netprioR}} object with fitted model
}
\description{
Fit \code{\linkS4class{netprioR}} model
}
\examples{
data(simulation)
np <- netprioR(networks = simulation$networks,
               phenotypes = simulation$phenotypes,
               labels = simulation$labels.obs,
               model.fit = FALSE)
summary(np)
np <- fit(np, nrestarts = 1, verbose = FALSE)
summary(np)
}
\author{
Fabian Schmich
}

