% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-functions.R
\name{laplacian}
\alias{laplacian}
\title{Graph Laplacian}
\usage{
laplacian(x, norm = c("none", "sym", "asym"))
}
\arguments{
\item{x}{Adjacency matrix}

\item{norm}{Type of normalisation}
}
\value{
Laplacian matrix
}
\description{
Compute the Laplacian matrix of a graph given its adjacency matrix
}
\author{
Fabian Schmich
}

