% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-methods.R
\docType{methods}
\name{weights}
\alias{ranks,netprioR-method}
\alias{weights}
\alias{weights,netprioR-method}
\title{Retrieve network weights and prioritisation ranks from \code{\linkS4class{netprioR}} objects}
\usage{
weights(object, ...)

\S4method{weights}{netprioR}(object)

\S4method{ranks}{netprioR}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{netprioR}} object}
}
\value{
Estimated network weights or prioritisation ranks
}
\description{
Retrieve network weights and prioritisation ranks from \code{\linkS4class{netprioR}} objects
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{ranks}}

\code{\link{weights}}
}

