\name{netprioR-package}
\alias{netprioR-package}
\alias{netprioR}
\docType{package}
\title{
\packageTitle{netprioR}
}
\description{
\packageDescription{netprioR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{netprioR}
\packageIndices{netprioR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{netprioR}

Maintainer: \packageMaintainer{netprioR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
