% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-package.R
\docType{package}
\name{netprioR-package}
\alias{netprioR-package}
\title{Package: netprioR}
\description{
This package provides a model for semi-supervised prioritisation of genes integrating network data, 
phenotypes and additional prior knowledge about TP and TN gene labels.
}
\author{
Fabian Schmich | Computational Biology Group, ETH Zurich | \email{fabian.schmich@bsse.ethz.ch}
}
\references{
Fabian Schmich et. al (2016).
}
\keyword{package}

