% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netprioR-simulation.R
\name{simulate_phenotype}
\alias{simulate_phenotype}
\title{Simulate phenotypes correlated to labels pivoted into two groups}
\usage{
simulate_phenotype(labels.true, meandiff, sd)
}
\arguments{
\item{labels.true}{Vector of labels}

\item{meandiff}{difference of means between positive and negative groups}

\item{sd}{Standard deviation of the phenotype}
}
\value{
Simulated phenotype
}
\description{
Simulate phenotypes correlated to labels pivoted into two groups
}
\examples{
data(simulation)
phenotypes <- simulate_phenotype(labels.true = simulation$labels.true, meandiff = 0.5, sd = 1)
}
\author{
Fabian Schmich
}

